+++ Changelog.md
@@ -10,6 +10,7 @@ This document tracks all notables changes to the Level Up app.
 
 - Add backup and restore data options
 - Add reset icons and reset settings options
+  - When resetting icons, all quest and quest history icons are cleared and default to the question mark icon
 - Add more icons
 
 ### Changed